Fixing p4 add folder_name
p4 uses `/...` as a wildcard to indicate "walk the tree and find all the
children of this" for all commands but add.  `p4 add folder_name/...`
fails, but that's what this bundle's Add to Depot command used to do.
Instead, we're now using Python's built-in os.walk to expand the file
list before we pass it to p4python for processing.  This should ensure
consistent (and expected) functionality across all commands.